A sample run on a real time Data
The file Names.csv is the one to be made by the receptionist. The file Patients.csv is the one from the google form. And the Results.csv is the merged dataframe with the output field ad prediction. 
A "No" means that the person will show up at the appointment and a "Yes" means that a person will now show up at the appointment.